requirements.txt
tensorflow==2.16.1

fastapi==0.111.0
uvicorn==0.30.1

numpy==1.26.4
matplotlib==3.9.0
scikit-learn==1.5.1

pillow==10.4.0
python-multipart==0.0.9

gdown==6.0.0